Add PriceInput component
