+++ src/modules/search/components/form/filter/PriceInput.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import { TextInput } from 'hds-react';
+import { FilterItem, FilterName } from '../../../../../types/common';
+import useFilters from '../../../../../hooks/useFilters';
+import styles from './PriceInput.module.scss';
+
+type Props = {
+  name: FilterName;
+  isWrapped?: boolean;
+  label: string;
+  items: FilterItem[];
+};
+
+const PriceInput = ({ name, label, isWrapped, items }: Props) => {
+  const { getFilter, setFilter } = useFilters();
+
+  const value = getFilter(name) || '';
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setFilter(name, event.target.value || '');
+  };
+
+  return (
+    <TextInput
+      id={`${name}-${label}`}
+      label={label}
+      value={value}
+      onChange={handleChange}
+      className={styles.priceInput}
+      style={isWrapped ? { padding: '14px' } : {}}
+      {...items[0]}
+    />
+  );
+};
+
+export default PriceInput;
